==> 01-Program-Years/2021/Project Code/dspg21oss/src/06_bert/01_bert_init_CZ.ipynb <==
# BERT on Repo Description

1. Construct a sentence corpus for each software type using labeled/manually validated repo descriptions
2. Calculate embeddingfor each corpus
3. Compare all repo description with each software type corpus using cosine-similarity score
    - It took about 1.5 hrs to run the embedding on repo data

Author: Cierra and Crystal

%reset

#pgadmin
import os
import psycopg2 as pg


#bert
from sentence_transformers import SentenceTransformer, util
import torch

import pandas as pd
import numpy as np

import re

import nltk
nltk.download("punkt")

from nltk import tokenize

import scipy

import datetime

# Embedding Model

#embedder = SentenceTransformer('paraphrase-MiniLM-L6-v2') #quicker model
embedder = SentenceTransformer('paraphrase-mpnet-base-v2') #most accurate, long run time

# Data

## I. Unlabelled Repo Data

repo_data = pd.read_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/clean_github_repos_157k.csv")
len(repo_data)

repo_data = repo_data[repo_data.description.notnull()]
len(repo_data)

np_repo_embedding = np.load('/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_description_embeddings.npy')

len(np_repo_embedding[0])

## II. Labelled Repo Data -- Software Type Corpus

data = pd.read_excel('/home/zz3hs/git/dspg21oss/data/dspg21oss/oss_software_labelled_blockchain_cz.xlsx') #import csv
#data = pd.read_excel('/home/zz3hs/git/dspg21oss/data/dspg21oss/oss_software_labelled_dataviz_cz.xlsx') #import csv

data

# software type
type_name =  "blockchain_label2"
#type_name =  "dataviz_label2"
#filter 500 validated repos that are labelled 1 (numeric)
corpus_type_i = data[data[type_name] ==1][["slug",type_name, "description_cleaneng", "readme"]]

corpus_type_i

# description sentence token
desc_tokens = corpus_type_i["description_cleaneng"].tolist()
desc_tokens

# readme sentence token

readmes =  corpus_type_i["readme"].tolist()

#tokenize.sent_tokenize(readme)
readmes_ls = []
for readme in readmes:
    readmes_ls.append(readme)
readmes_ls_filtered = [x for x in readmes_ls if x != "a"]

readmes_str = ''.join(readmes_ls_filtered) 
readmes_str

readme_tokens =tokenize.sent_tokenize(readmes_str)
readme_tokens

corpus_type_i_token = desc_tokens + readme_tokens
corpus_type_i_token

# Embedding of corpus

# embedding for the corpus
corpus_type_i_embeddings = embedder.encode(corpus_type_i_token, show_progress_bar=True) # embeddings


# Similarity Score

takes about 35 seconds to run

queries = np_repo_embedding

# pre-specified number of sentences
num_sentences = 10 #find 10 most similar sentences from the corpus

# init a result list for scores
result = []


t1 = datetime.datetime.now()
print("Start:", t1)

for query in queries: #compare each sentence in the abstract to the software type corpus
    
    # We use cosine-similarity and torch.topk to find the highest k scores
    cos_scores = util.pytorch_cos_sim(query, corpus_type_i_embeddings)[0]

    top_results = torch.topk(cos_scores, k=num_sentences)   #get the top k scores
    result.append(top_results.values.tolist()) #unlist the top result list
   
    #print 10 most similar entences from the corpus and their corresponding scores
    #print("\n\n======================\n\n")
    #print("Query:", query)
    #print("Results:", top_results)
    #print("\nTop k=10 most similar sentences in corpus:")
    #for score, idx in zip(top_results[0], top_results[1]):
    #    print(corpus_type_i_clean[idx], "(Score: {:.4f})".format(score))

t2 =  datetime.datetime.now()
print("Finished", len(result), "descriptions at", t2)
print("It took", t2-t1, "to run.")

#TODO: THIS WAY, similarity_score IS SAVED AS A STRING, NEED TO FIGURE OUT HOW TO SAVE AS A LIST
#save the similarity score as a variable of the original repo data
repo_data["similarity_score"] = result

repo_data

import statistics #calculate mean and others

sentence_score_ls = repo_data["similarity_score"]

mean_sentence_score= []
max_sentence_score = []
median_sentence_score = []
for sentence_score in sentence_score_ls:
    mean_sentence_score.append(statistics.mean(sentence_score))
    max_sentence_score.append(max(sentence_score))
    median_sentence_score.append(statistics.median(sentence_score))

repo_data["mean_sentence_score"]=mean_sentence_score
repo_data["max_sentence_score"]=max_sentence_score
repo_data["median_sentence_score"]=median_sentence_score


repo_data

#save csv
#repo_data.to_csv(r'/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_data_blockchain_score.csv', index = False)   
#repo_data.to_csv(r'/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_data_dataviz_score.csv', index = False)   


# Similarity Score Analysis

from scipy import stats
from scipy.stats import skew
import statistics #calculate mean and others

#read in data
repo_data = pd.read_csv(r'/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_data_python_score.csv')   


repo_data

#score is in a string, convert to a list, also make sure the numbers are float
score_ls = repo_data["similarity_score"]

score_ls_float = []
for sentence_score in score_ls:
    sentence_score = str(sentence_score)[1:-1]
    sentence_score = sentence_score.split(",")
    item_float= []
    for item in sentence_score:
        item_float.append(float(item))
    score_ls_float.append(item_float)

    
repo_data["similarity_score_float"] = score_ls_float

#check scores are in a list
repo_data["similarity_score_float"][0][0]

# get score statistics
score_ls = repo_data["similarity_score_float"]

mean_score= []
range_score = []
max_score = []
median_score = []
skewness_score = []
for sentence_score in score_ls:
    mean_score.append(statistics.mean(sentence_score))
    range_score.append(max(sentence_score)- min(sentence_score))
    max_score.append(max(sentence_score))
    median_score.append(statistics.median(sentence_score))
    skewness_score.append(stats.skew(sentence_score))
repo_data["mean_score"]=mean_score
repo_data["range_score"]=range_score
repo_data["max_score"]=max_score
repo_data["median_score"]=median_score
repo_data["skewness_score"]=skewness_score

repo_data
